+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### vNext
+
+- Fix geyser rate modification.
+
 ### 3.5.0
 
 - Change geyser game object type (and artwork) when changing emit element.